% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diapha_pvb}
\alias{diapha_pvb}
\title{Diaphanography test data set}
\format{
A data frame with 900 rows and three variables:
\describe{
  \item{disease:}{Breast cancer, \eqn{disease}: Binary, 1 = Yes, 0 = No}
  \item{test:}{Diaphanography, \eqn{test}: Binary, 1 = Positive, 0 = Negative}
  \item{verified:}{Verified, \eqn{verified}: Binary, 1 = Yes, 0 = No}
}
}
\source{
\enumerate{
  \item{Marshall, V., Williams, D. C., & Smith, K. D. (1984). Diaphanography as a means of detecting breast cancer. Radiology, 150(2), 339–343.}
}
}
\usage{
diapha_pvb
}
\description{
Diaphanography test is a noninvasive method (diagnostic test) of breast examination by transillumination using visible or infrared light to detect the presence of breast cancer. The test was performed on 900 patients. Only 88 patients were verified by breast tissue biopsy for histological examination (gold standard test). The percentage of unverified patients is 90.2\%.
}
\keyword{datasets}
