% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hepatic_pvb}
\alias{hepatic_pvb}
\title{Hepatic scintigraphy test data set}
\format{
A data frame with 650 rows and three variables:
\describe{
  \item{disease:}{Liver cancer, \eqn{disease}: Binary, 1 = Yes, 0 = No}
  \item{test:}{Hepatic scintigraphy, \eqn{test}: Binary, 1 = Positive, 0 = Negative}
  \item{verified:}{Verified, \eqn{verified}: Binary, 1 = Yes, 0 = No}
}
}
\source{
\enumerate{
  \item{Drum, D. E., & Christacopoulos, J. S. (1972). Hepatic scintigraphy in clinical decision making. Journal of Nuclear Medicine, 13(12), 908–915.}
}
}
\usage{
hepatic_pvb
}
\description{
The data set pertains to hepatic scintigraphy, a diagnostic imaging technique used for detecting liver cancer. The test was performed on 650 patients, where 344 patients were verified by liver pathological examination (gold standard test). The percentage of unverified patients is 47.1\%.
}
\keyword{datasets}
